requirements.txt
boto3
python-dotenv
fastapi
uvicorn
pydantic
circuitbreaker
